agent/requirements.txt
# === CORE (Always needed) ===
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0

# === JOB SCRAPING ===
python-jobspy>=1.1.0

# === BROWSER AUTOMATION (Optional - for auto-submission) ===
# Uncomment these for full autonomous mode:
# browser-use>=0.1.0
# playwright>=1.40.0
# langchain-openai>=0.1.0
# After installing, run: playwright install chromium

# === GITHUB INTEGRATION (Optional) ===
PyGithub>=2.0.0
